semester-3/python-iii/flask-231129/pyproject.toml
[project]
name = "flask-231129"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "pan93412", email = "pan93412@gmail.com" }
]
dependencies = [
    "flask>=3.0.0",
    "sqlalchemy>=2.0.23",
    "argon2-cffi>=23.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.11.0",
]

[tool.hatch.metadata]
allow-direct-references = true
